requirements/requirements.txt
aiohttp==0.22.0
chardet==2.3.0
contextlib2==0.5.3
coverage==4.1
gevent==1.1.1
greenlet==0.4.10
grequests==0.3.0
httpretty==0.8.14
iniherit==0.3.4
Momoko==2.2.3
multidict==1.1.0
psycopg2==2.6.2
py==1.4.31
pytest==2.9.2
pytest-cov==2.3.0
pytest-tornado==0.4.5
python-slugify==1.2.0
raven==5.23.0
requests==2.10.0
six==1.10.0
tornado==4.4
Unidecode==0.4.19
yoyo-migrations==5.0.3
